move bingo_tile to widgets folder
